requirements.txt
numpy==1.21.6
scipy==1.7.3
scikit-learn==1.0.2
networkx==2.6.3
pandas==1.3.5
matplotlib==3.7.0
tqdm==4.65.0
dgl==1.1.3
torchdata==0.6.0
torch-geometric==2.3.1
wandb